# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "an.nguyenbku1999@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/session/add": {
            "post": {
                "tags": [
                    "Session"
                ],
                "description": "add Session",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter name Session:",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/session/edit": {
            "post": {
                "tags": [
                    "Session"
                ],
                "description": "edit Session",
                "parameters": [
                    {
                        "name": "SessionId",
                        "in": "query",
                        "description": "Enter session ID:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter name session:",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/session/delete": {
            "post": {
                "tags": [
                    "Session"
                ],
                "description": "delete Session",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "SessionId",
                        "in": "query",
                        "description": "Enter session ID:",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/session/showSessionsList": {
            "get": {
                "tags": [
                    "Session"
                ],
                "description": "show Sessions List",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
...
    },
...
}